[wip] split UserInterface into smaller files
