+++ scripts/lib/highnoon/SessionState.js
